Pin scroll view to bottom/trailing edges (#130)
If the content size changes while the scroll view is scrolled all the
way to the bottom/trailing edges, the scroll offset will be
automatically adjusted, "pinning" that edge.


https://github.com/user-attachments/assets/f503c186-02aa-4692-951f-c93253b6a5ec

The animation isn't perfect if an element in the middle changes size
while pinned, but scroll already has a couple other minor animation
timing inconsistencies that can probably be fixed in a future iteration.
Likely requires more extensive rework of scroll internals to fix.